fix: An error that ends when the floating result window presses any keyother than "q"(플로팅 결과 창이 "q"가 아닌 아무키를 눌렸을 때도 종료되는 오류)
When I compiled and executed, I found an error that the floating window disappeared when I pressed any key other than "q" in the floating window.

First of all, the floating screen is Neovim's built-in terminal session. For programs that only output, you don't have to use the built-in terminal session, but for programs that require input, you have to use the session of the built-in terminal to put the input in and run the program. I found that if you run the program normally in the built-in terminal, [Process limited 0] state, and when you press any key in this state, the floating screen turns off.

In order to solve this problem, when the program was terminated normally, that is, when it was in the [Process limited 0] state, it was switched back to the vim normal mode so that the output log could be checked by moving the cursor, and the floating screen could be terminated by pressing "q" or entering the ":q" vim's quit command.

In addition, I used the api of nvim for the type of floating screen and adjusted the size and position to make it look like a border, so I modified it to distinguish between the background screen and the floating screen better.

[Korean]:
컴파일 및 실행을 했을 때 뜨는 플로팅 창에서 "q"가 아닌 아무 키를 눌렸을 때도 플로팅 창이 없어지는 오류를 발견했습니다.

우선 플로팅 화면은 neovim의 내장 터미널 세션입니다. 단지 출력만 해주는 프로그램 같은 경우 내장 터미널 세션을 쓰지 않아도 되지만, 입력을 필요로 하는 프로그램 같은 경우 내장 터미널의 세션을 써야지만 입력을 넣고 프로그램을 실행할 수 있습니다. 내장 터미널에서 프로그램을 정상적으로 수행하면 [Process exited 0] 상태가 되며, 이 상태일 때 아무 키를 누르게 되면 플로팅 화면이 꺼지는 것을 발견했습니다.

해결하기 위해서 프로그램이 정상적으로 수행된 종료되었을 때, 즉 [Process exited 0] 상태가 되면 다시 vim normal모드로 전환되도록 하여 출력 로그를 커서를 움직여 확인할 수 있게 하였고, "q"를 눌리거나, ":q" vim의 quit 명령어를 입력하여 플로팅 화면을 종료할 수 있도록 하였습니다.

추가적으로 플로팅 화면의 종류를 nvim의 api를 이용했고 크기와 위치를 조정하여 border가 생긴 것 처럼 작동하기 때문에 백그라운드 화면과 플로팅 화면을 좀 더 잘 구분할 수 있도록 수정하였습니다.