fix: Bug where compilation error output does not appear in the floating window
In the case of a language that compiles first, such as java, c, and c++ (javac, gcc, g++..), when compilation fails, the compilation error log does not appear on the floating screen, but the previously successful compilation file is just executed and the result is displayed on the floating screen. In the case of compilation error, the compilation error log was output on the floating screen, and when compilation was successful, the program execution result was modified to appear on the floating screen.

For commands that compile and execute at once, such as javascript, python(node, python3), is fine.

Currently, only the c++, java, python, and javascript languages are set in my configuration file. If you want to add another language, please leave it in issue and let me know, or if you work directly to modify the configuration file and leave a pull request, I will check and apply it

[Korean]:
java, c, c++ 같은 컴파일을 먼저 하고 난 후 실행하는 언어 같은 경우(javac, gcc, g++..) 컴파일 실패 시 컴파일 에러 로그가 플로팅 화면에 뜨지 않고 이전 컴파일 성공했던 파일을 그냥 실행시켜 결과가 플로팅 화면에 뜨는 버그를 발견하였다. 컴파일 에러 시 컴파일 에러 로그를 플로팅 화면으로 출력되도록 하였고, 컴파일 성공 시 프로그램 수행 결과를 플로팅 화면으로 뜨도록 수정하였다.

javascript, python(node, python3)과 같이 컴파일 및 실행을 한 번에 하는 명령어 같은 경우는 괜찮다.

현재 저의 설정파일에는 c++, java, python, javascript 언어만 설정되어있다. 혹시 다른 언어를 추가하고싶다면 issue에 남겨 저한테 알려주시거나 혹은 직접 fork하여 설정파일을 수정하여 pull request를 남기면 확인 후 적용하겠음